requirements.txt
mmcv-full==1.3.14
mmsegmentation==0.17.0
numpy==1.21.1
prettytable==2.1.0
segmentation-models-pytorch==0.2.0
torch==1.9.1+cu102
torchvision==0.10.1+cu102
